Update 9891 Project Code.R
Removed a section from the loop which formatted the data set. This was no longer needed since all formatting was done at the top Data Preparation section. Updated the indexing for the full models. Removed the single lambda value that was passed to glmnet() because supplying a single value led the model to be unable to converge. Added parallel to all cross validation calls to speed up computation.